chore: do not render credentials container if ITW is not valid
