+++ ts/features/wallet/components/WalletCardsContainer.tsx
@@ -1,41 +1,27 @@
 import { IOStyles, ListItemHeader } from "@pagopa/io-app-design-system";
-import { useFocusEffect } from "@react-navigation/native";
-import { useCallback, useMemo } from "react";
+import { useMemo } from "react";
 import { View } from "react-native";
 import Animated, { LinearTransition } from "react-native-reanimated";
 import { useDebugInfo } from "../../../hooks/useDebugInfo";
 import I18n from "../../../i18n";
-import { useIONavigation } from "../../../navigation/params/AppParamsList";
 import { useIOSelector } from "../../../store/hooks";
-import { useIOBottomSheetAutoresizableModal } from "../../../utils/hooks/bottomSheet";
-import {
-  ItwEidInfoBottomSheetContent,
-  ItwEidInfoBottomSheetTitle
-} from "../../itwallet/common/components/ItwEidInfoBottomSheetContent";
-import { ItwEidLifecycleAlert } from "../../itwallet/common/components/ItwEidLifecycleAlert";
-import { ItwFeedbackBanner } from "../../itwallet/common/components/ItwFeedbackBanner";
-import { ItwWalletReadyBanner } from "../../itwallet/common/components/ItwWalletReadyBanner";
+import { ItwWalletNotAvailableBanner } from "../../itwallet/common/components/ItwWalletNotAvailableBanner";
 import { ItwDiscoveryBannerStandalone } from "../../itwallet/common/components/discoveryBanner/ItwDiscoveryBannerStandalone";
-import { itwCredentialsEidStatusSelector } from "../../itwallet/credentials/store/selectors";
-import { itwLifecycleIsValidSelector } from "../../itwallet/lifecycle/store/selectors";
+import { ItwWalletCardsContainer } from "../../itwallet/wallet/components/ItwWalletCardsContainer";
 import { useItwWalletInstanceRevocationAlert } from "../../itwallet/walletInstance/hook/useItwWalletInstanceRevocationAlert";
 import {
   isWalletEmptySelector,
   selectIsWalletLoading,
-  selectWalletCardsByCategory,
   selectWalletCategories,
   selectWalletOtherCards,
   shouldRenderItwCardsContainerSelector,
   shouldRenderWalletEmptyStateSelector
 } from "../store/selectors";
-import { ItwWalletNotAvailableBanner } from "../../itwallet/common/components/ItwWalletNotAvailableBanner";
 import { withWalletCategoryFilter } from "../utils";
 import { WalletCardSkeleton } from "./WalletCardSkeleton";
 import { WalletCardsCategoryContainer } from "./WalletCardsCategoryContainer";
 import { WalletCardsCategoryRetryErrorBanner } from "./WalletCardsCategoryRetryErrorBanner";
 import { WalletEmptyScreenContent } from "./WalletEmptyScreenContent";
-
-const EID_INFO_BOTTOM_PADDING = 128;
 
 /**
  * A component which renders the wallet cards container
@@ -100,91 +86,6 @@ const WalletCardsContainerSkeleton = () => (
     <WalletCardSkeleton testID="walletCardSkeletonTestID_3" cardProps={{}} />
   </>
 );
-
-/**
- * Card container for the ITW credentials
- */
-const ItwWalletCardsContainer = withWalletCategoryFilter("itw", () => {
-  const navigation = useIONavigation();
-  const cards = useIOSelector(state =>
-    selectWalletCardsByCategory(state, "itw")
-  );
-  const isItwValid = useIOSelector(itwLifecycleIsValidSelector);
-  const eidStatus = useIOSelector(itwCredentialsEidStatusSelector);
-
-  const isEidExpired = eidStatus === "jwtExpired";
-
-  useDebugInfo({
-    itw: {
-      isItwValid,
-      eidStatus,
-      cards
-    }
-  });
-
-  const eidInfoBottomSheet = useIOBottomSheetAutoresizableModal(
-    {
-      title: <ItwEidInfoBottomSheetTitle isExpired={isEidExpired} />,
-      // Navigation does not seem to work when the bottom sheet's component is not inline
-      component: <ItwEidInfoBottomSheetContent navigation={navigation} />
-    },
-    EID_INFO_BOTTOM_PADDING
-  );
-
-  useFocusEffect(
-    useCallback(
-      // Automatically dismiss the bottom sheet when focus is lost
-      () => eidInfoBottomSheet.dismiss,
-      [eidInfoBottomSheet.dismiss]
-    )
-  );
-
-  const sectionHeader = useMemo((): ListItemHeader | undefined => {
-    if (!isItwValid) {
-      return undefined;
-    }
-    return {
-      testID: "walletCardsCategoryItwHeaderTestID",
-      iconName: "legalValue",
-      iconColor: isEidExpired ? "grey-300" : "blueIO-500",
-      label: I18n.t("features.wallet.cards.categories.itw"),
-      endElement: {
-        type: "buttonLink",
-        componentProps: {
-          accessibilityLabel: I18n.t(
-            "features.itWallet.presentation.bottomSheets.eidInfo.triggerLabel"
-          ),
-          label: I18n.t(
-            "features.itWallet.presentation.bottomSheets.eidInfo.triggerLabel"
-          ),
-          onPress: eidInfoBottomSheet.present,
-          testID: "walletCardsCategoryItwActiveBadgeTestID"
-        }
-      }
-    };
-  }, [isItwValid, isEidExpired, eidInfoBottomSheet.present]);
-
-  return (
-    <>
-      <WalletCardsCategoryContainer
-        key={`cards_category_itw`}
-        testID={`itwWalletCardsContainerTestID`}
-        cards={cards}
-        header={sectionHeader}
-        topElement={
-          <>
-            <ItwWalletReadyBanner />
-            <ItwEidLifecycleAlert
-              lifecycleStatus={["jwtExpiring", "jwtExpired"]}
-            />
-          </>
-        }
-        bottomElement={<ItwFeedbackBanner />}
-      />
-      {isItwValid && eidInfoBottomSheet.bottomSheet}
-    </>
-  );
-});
 
 /**
  * Card container for the other cards (payments, bonus, etc.)
